+++ packages/connectivity/src/scp-cf/destination/destination-accessor-types.ts
@@ -1,4 +1,8 @@
-import { Destination } from './destination-service-types';
+import { VerifyJwtOptions } from '../jwt';
+import {
+  Destination,
+  DestinationRetrievalOptions
+} from './destination-service-types';
 import { DestinationSelectionStrategy } from './destination-selection-strategies';
 
 /**
@@ -45,4 +49,29 @@ export interface DestinationAccessorOptions {
    * So be careful that the used value is not manipulated and breaks the tenant isolation of your application.
    */
   iss?: string;
+}
+
+export type DestinationOptions = DestinationAccessorOptions &
+  DestinationRetrievalOptions &
+  VerifyJwtOptions;
+
+/**
+ * Declaration of a destination to be retrieved from an environment variable or from the destination service on SAP Business Technology Platform, including all DestinationOptions.
+ *
+ * Use an object of this interface to specify which destination shall be used when executing a request.
+ * The destination will be retrieved via its [[DestinationFetchOptions.destinationName]] according to the following algorithm:
+ * 1. If a destination of this [[DestinationFetchOptions.destinationName]] is defined in the environment variable `destinations` (if available), it will be converted into a [[Destination]] and used for the request.
+ * 2. Otherwise, the destination service on SAP Business Technology Platform is queried for a destination with the given [[DestinationFetchOptions.destinationName]], using the access token provided as value of property [[jwt]].
+ * Additionally, you can set [[DestinationOptions]] for objects of this interface.
+ */
+export interface DestinationFetchOptions extends DestinationOptions {
+  /**
+   * Name of the destination to retrieve, mandatory.
+   */
+  destinationName: string;
+
+  /**
+   * An access token for the XSUAA service on SAP Business Technology Platform, provided as a JSON Web Token, only mandatory when destination shall be retrieved from destination service on SAP Business Technology Platform.
+   */
+  jwt?: string;
 }
