+++ packages/connectivity/src/scp-cf/destination/destination-service-types.ts
@@ -1,8 +1,6 @@
 import { CachingOptions } from '../cache';
 import { ProxyConfiguration } from '../connectivity-service-types';
-import { VerifyJwtOptions } from '../jwt';
 import { ResilienceOptions } from '../resilience-options';
-import { DestinationAccessorOptions } from './destination-accessor-types';
 
 /**
  * A resolved destination containing information needed to execute requests, such as the system URL.
@@ -189,47 +187,10 @@ export interface DestinationCertificate {
   type: string;
 }
 
-export type DestinationOptions = DestinationAccessorOptions &
-  DestinationRetrievalOptions &
-  VerifyJwtOptions;
-
-/**
- * Declaration of a destination to be retrieved from an environment variable or from the destination service on SAP Business Technology Platform, including all DestinationOptions.
- *
- * Use an object of this interface to specify which destination shall be used when executing a request.
- * The destination will be retrieved via its [[DestinationFetchOptions.destinationName]] according to the following algorithm:
- * 1. If a destination of this [[DestinationFetchOptions.destinationName]] is defined in the environment variable `destinations` (if available), it will be converted into a [[Destination]] and used for the request.
- * 2. Otherwise, the destination service on SAP Business Technology Platform is queried for a destination with the given [[DestinationFetchOptions.destinationName]], using the access token provided as value of property [[jwt]].
- * Additionally, you can set [[DestinationOptions]] for objects of this interface.
- */
-export interface DestinationFetchOptions extends DestinationOptions {
-  /**
-   * Name of the destination to retrieve, mandatory.
-   */
-  destinationName: string;
-
-  /**
-   * An access token for the XSUAA service on SAP Business Technology Platform, provided as a JSON Web Token, only mandatory when destination shall be retrieved from destination service on SAP Business Technology Platform.
-   */
-  jwt?: string;
-}
-
 /**
  * Options to use while fetching destinations. Encompasses both [[DestinationCachingOptions]] and [[ResilienceOptions]] interfaces.
  */
 export type DestinationRetrievalOptions = CachingOptions & ResilienceOptions;
-
-/**
- * Typeguard to find if object is DestinationNameAndJwt.
- * @param destination - Destination to be checked
- * @returns boolean
- * @internal
- */
-export function isDestinationFetchOptions(
-  destination: any
-): destination is DestinationFetchOptions {
-  return destination.destinationName !== undefined;
-}
 
 /**
  * Typeguard to find if object is DestinationNameAndJwt.
